Remove palette from color scale default config
This is to prevent attempting to set a default when it's already there. Color scale palette can never be undefined though because there is special logic in the getter. That also handles the case when palette is missing from config or default config. Strange though that if (!config.hasOwnProperty("palette")) doesn't work in utils. It's still entering the IF statement, it shouldn't.

Fix for https://github.com/Gapminder/tools-page/issues/123