Frame specific autoconfig, new solver api & bugfixes
Frame now solves concepts to be preferable in space, preferably time typed.

New solver API:
Replace autoconfig with filter
Move selectMethod to sibling of filter
I.e. if concept/space is object, it automatically is autoconfig
May interfere in future with object properties we want to autoconfig, but this makes current config terser. We'll worry about it when it shows up.