Enable filter.set to set payload from config reference instead of explicitly
Used in find dialog in bubble chart. Find just sets selected filter. But selected filter is a reference to trail filter. Trail filter wants a special payload (start year). It can get this payload through trail encoding. This is two steps up in the model tree from the filter. Thus set payload in filter to be a relative reference from the filter.

This can't be an absolute reference because we don't know encoding name when setting the default reference (it's often "trail", but may not be).

We need the selected filter reference to be a reference to the trail filter _model_, not _config_, as a _config_ just builds a new filter model whose relative reference goes to select encoding (which doesn't have access to trail payload). I.e. the filter.set should go to trail filter model, not selected filter model, thus we need a model reference.

Now any reference to filter will be model, but we should probably let the user decide.