Don't use a reference in solving marker config to prevent loop
Loop example:
get size -> get marker solution -> get order solution -> resolve order concept reference -> get size

Now reference is resolved only after configSolution has been found (and encodings with refs were skipped in this process).
I.e. the solution for referenced concepts is just: your reference.

Now, if this reference was not pointing to a concept used in same solver, there would be no loop and you could've resolved the reference to take it into account when solving. But this is an edge case. Worst case it leads to a two concepts being the same. If later it leads to worse cases, may need to make this smarter.