Make references stateful to enable simple legend state checking
Stateful references are tricky: they open up possibility for a lot more dependency loops. E.g. in marker encodings that refer to eachother. Most are solved by changing simple properties to getters, so that computeds are only calculated if they are actually read.
This helped both in the marker-encoding-self reference and for making sure referenceState was solved before loading metadata and autosolving.

One lingering problem is that combining states now will not return 'rejected' if one of the promises is rejected. For that it would always have to check all states. But that would trigger all getters. I.e. there's sequential and parallel combine states, we now moved from parallel to sequential. Probably should just have two separate utils for this.