Add "interval" property to frame model to be able to set it in config...
when data is by days or weeks etc. The interval can not always be identified via the name of time concept, i.e. in case of "time". Previously the interval would always default to "year", now it's configurable. Subsequently i had to raise the logic of choosing the interval from utils to frame.