Feature: make encodings backfill data when enc.concept is an entity set
and the domain of that entity set is among the space.

Before we would backfill the data for world_4region.world_4region via conceptInSpace() feature, but not for geo.world_4region. Now we can also handle this other case too. 

This allows for example to resolve colors of world_4regions when space is [geo] and color concept is world_4region.