v1.35.1: Fix user-configurable default palette => branded constant color
Problems: First, if users sets partial defaultPalette, some concepts may trigger error. This is fixed by merging configs and defaults instead of choosing one or the other. This is fixed in get defaultPalettes(). Second, switching of indicators would reset not only the runtime palette modifications, like it should, but also user-configurable defaults. This is fixed in encoding.setWhich