v1.37.0: Complete refactor of filter "limited structure" helpers
The main idea is that the limited structure shape changes to allow for more flexible filters: every entity level (called isness because of is--level kind of notation) gets its own section in dim.$or array. The subtractive section now takes the same shape using $nor array. Thus we can use the same operations in symmetry for adding and removing items using limited structure.